// cosmowcl/src/app/shared/models/api/cosmoauthorization/creaParametroDiSistemaRequest.d.ts
/*
 * Copyright CSI-Piemonte - 2023
 * SPDX-License-Identifier: GPL-3.0-or-later
 */
/**
 * configurazione
 *  API di configurazione generale e di preferenze utente
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 */
export interface CreaParametroDiSistemaRequest { 
    chiave: string;
    valore: string;
    descrizione?: string;
}